8307495: Specialize atomic bitset functions for aix-ppc
